refactor(auth): single-source Better Auth config + split package exports (#45)
Collapse the type-inference-only betterAuth() stub and the auth-service
runtime instance into one source of truth in @reloop/auth/server. Split
exports so client/types stay free of database, Redis, and bus deps.

- Move real betterAuth config, secondary storage, signup-invite helpers
  into packages/auth/src/server/
- Add @reloop/auth/types; client derives AuthInstance from types only
- Thin be-auth to re-export the package runtime instance
- Export-isolation tests + CI gate; #44 characterization still green
