+++ .github/workflows/be-auth.yml
@@ -5,12 +5,14 @@ on:
     branches: [ main ]
     paths:
       - 'apps/backend/auth/**'
+      - 'packages/auth/**'
       - 'docker-compose.setup.yml'
       - '.github/workflows/be-auth.yml'
   pull_request:
     branches: [ main ]
     paths:
       - 'apps/backend/auth/**'
+      - 'packages/auth/**'
       - 'docker-compose.setup.yml'
       - '.github/workflows/be-auth.yml'
   workflow_dispatch:
@@ -33,6 +35,10 @@ jobs:
 
       - name: Install dependencies
         run: bun install --frozen-lockfile
+
+      - name: Run @reloop/auth export isolation tests
+        run: bun run test
+        working-directory: packages/auth
 
       - name: Run auth characterization tests
         run: bun run test
