+++ apps/backend/auth/src/lib/auth.ts
@@ -1,452 +1,4 @@
-import { apiKey } from "@better-auth/api-key";
-import { ac, orgRoles } from "@reloop/auth/permissions";
-import { BusEvent, bus } from "@reloop/bus";
-import { db } from "@reloop/db/client";
-import * as schema from "@reloop/db/schema";
-import { betterAuth } from "better-auth";
-import { drizzleAdapter } from "better-auth/adapters/drizzle";
-import { APIError, createAuthMiddleware } from "better-auth/api";
-import {
-	admin,
-	bearer,
-	emailOTP,
-	jwt,
-	lastLoginMethod,
-	openAPI,
-	organization,
-} from "better-auth/plugins";
-import { eq } from "drizzle-orm";
-import { log } from "evlog";
-import { authConfig } from "../auth.config";
-import {
-	DEFAULT_USER_ROLE,
-	PLATFORM_ADMIN_ROLE,
-	platformAc,
-	platformRoles,
-} from "./platform-roles";
-import { redis } from "./redis";
-import {
-	canCreateAccount,
-	findValidSignupInvite,
-	getSignupInviteCodeFromRequest,
-	markSignupInviteUsed,
-	normalizeEmail,
-	userExistsByEmail,
-} from "./signup-invite";
-
-export const auth = betterAuth({
-	database: drizzleAdapter(db, {
-		provider: "pg",
-		schema: schema,
-	}),
-	user: {
-		additionalFields: {
-			activeOrganizationId: {
-				type: "string",
-				required: false,
-				input: true,
-			},
-			mode: {
-				type: "string",
-				required: false,
-				input: true,
-				defaultValue: "dev",
-			},
-		},
-	},
-	secondaryStorage: {
-		get: async (key) => {
-			return await redis.get(key);
-		},
-		set: async (key, value, ttl) => {
-			if (ttl) await redis.set(key, value, ttl);
-			else await redis.set(key, value);
-		},
-		delete: async (key) => {
-			await redis.delete(key);
-		},
-	},
-	databaseHooks: {
-		user: {
-			create: {
-				before: async (user, context) => {
-					if (!authConfig.REQUIRE_SIGNUP_INVITE) return;
-
-					const email = normalizeEmail(user.email);
-					const code = context?.headers
-						? getSignupInviteCodeFromRequest(context.headers)
-						: null;
-					let access = await canCreateAccount({ email, code });
-					if (!access.allowed && code) {
-						access = await canCreateAccount({ email });
-					}
-
-					if (!access.allowed) {
-						throw new APIError("FORBIDDEN", {
-							message: "A valid signup invite is required to create an account",
-						});
-					}
-
-					if (access.signupInvite) {
-						await redis.set(
-							`signup_invite:pending:${email}`,
-							access.signupInvite.id,
-							60 * 30,
-						);
-					}
-				},
-				after: async (user) => {
-					if (!authConfig.REQUIRE_SIGNUP_INVITE) return;
-
-					const email = normalizeEmail(user.email);
-					const inviteId = await redis.get<string>(
-						`signup_invite:pending:${email}`,
-					);
-					if (inviteId) {
-						await markSignupInviteUsed({
-							inviteId,
-							userId: user.id,
-						});
-						await redis.delete(`signup_invite:pending:${email}`);
-						return;
-					}
-
-					const invite = await findValidSignupInvite({ email });
-					if (invite) {
-						await markSignupInviteUsed({
-							inviteId: invite.id,
-							userId: user.id,
-						});
-					}
-				},
-			},
-		},
-	},
-	hooks: {
-		after: createAuthMiddleware(async (ctx) => {
-			const { path, context } = ctx;
-			log.info({ message: String(ctx.path) });
-			// 🔐 User registered
-			if (path === "/sign-up/email-otp") {
-				const newSession = context.newSession;
-				if (newSession) {
-					log.info({
-						...{ data: newSession.user },
-						message: "🔐 User registered:",
-					});
-					await bus.publish(
-						BusEvent.USER_CREATED,
-						{
-							id: newSession.user.id,
-							email: newSession.user.email,
-							name: newSession.user.name || undefined,
-						},
-						{ msgId: `user_created:${newSession.user.email}` },
-					);
-				}
-			}
-
-			// 🔓 User signed in
-			if (
-				path === "/sign-in/email-otp" ||
-				path === "/callback/google" ||
-				path === "/callback/github"
-			) {
-				const data = context.newSession;
-				if (data) {
-					const { session, user } = data;
-					log.info({ ...{ data: user.email }, message: "🔓 User signed in:" });
-					// Use a 1-minute bucket for sign-in deduplication
-					const bucket = Math.floor(Date.now() / 60000);
-					await bus.publish(
-						BusEvent.SIGNIN_DETECTED,
-						{
-							email: user.email,
-							fullName: user.name || "User",
-							browser: session.userAgent || "Unknown Browser",
-							os: "Unknown OS",
-							ip: session.ipAddress || "0.0.0.0",
-							location: "Unknown Location",
-						},
-						{ msgId: `signin_detected:${user.email}:${bucket}` },
-					);
-				}
-			}
-		}),
-	},
-	basePath: "/api/auth/v1",
-	telemetry: { enabled: false },
-	emailAndPassword: {
-		enabled: true,
-		autoSignIn: true,
-		disableSignUp: authConfig.DISABLE_SIGNUP === "true",
-	},
-	socialProviders: {
-		google: {
-			clientId: authConfig.GOOGLE_CLIENT_ID as string,
-			clientSecret: authConfig.GOOGLE_CLIENT_SECRET as string,
-		},
-		github: {
-			clientId: authConfig.GITHUB_CLIENT_ID as string,
-			clientSecret: authConfig.GITHUB_CLIENT_SECRET as string,
-		},
-	},
-	secret: authConfig.BETTER_AUTH_SECRET,
-	session: {
-		expiresIn: 60 * 60 * 24 * 7,
-		updateAge: 60 * 60 * 24,
-	},
-	trustedOrigins: ["*"],
-	plugins: [
-		jwt(),
-		bearer(),
-		admin({
-			defaultRole: DEFAULT_USER_ROLE,
-			adminRoles: [PLATFORM_ADMIN_ROLE],
-			ac: platformAc,
-			roles: platformRoles,
-		}),
-		apiKey({ defaultPrefix: "rl" }),
-		lastLoginMethod({
-			cookieName: "better-auth.last_used_login_method",
-			maxAge: 60 * 60 * 24 * 30,
-			customResolveMethod: (ctx) => {
-				if (ctx.path.includes("/oauth/callback/google")) {
-					return "google";
-				}
-				if (ctx.path.includes("/oauth/callback/github")) {
-					return "github";
-				}
-				if (ctx.path.includes("/sign-up/email")) {
-					return "email";
-				}
-				if (ctx.path.includes("/sign-in/email")) {
-					return "email";
-				}
-				return null;
-			},
-		}),
-		emailOTP({
-			expiresIn: 60 * 15,
-			allowedAttempts: 3,
-			async sendVerificationOTP({ email, otp, type }) {
-				if (
-					authConfig.REQUIRE_SIGNUP_INVITE &&
-					(type === "sign-in" || type === "email-verification")
-				) {
-					const exists = await userExistsByEmail(email);
-					if (!exists) {
-						const access = await canCreateAccount({ email });
-						if (!access.allowed) {
-							throw new APIError("FORBIDDEN", {
-								message:
-									"A valid signup invite is required to create an account",
-							});
-						}
-					}
-				}
-
-				log.info("server", `Sending OTP (${type}) to: ${email} (OTP: ${otp})`);
-				if (authConfig.DEFAULT_OTP && authConfig.NODE_ENV !== "development")
-					return;
-				try {
-					await bus.publish(
-						BusEvent.OTP_REQUESTED,
-						{ email, otp, type },
-						{ msgId: `otp_requested:${email}:${otp}` },
-					);
-					log.info("server", `OTP bus event published for ${email} (${type})`);
-				} catch (error) {
-					log.error({
-						...{ data: error },
-						message: "Failed to publish OTP event:",
-					});
-					// If we are not using a default OTP, we must throw to notify the user
-					if (!authConfig.DEFAULT_OTP) {
-						throw new Error("Failed to send OTP email");
-					}
-				}
-			},
-			generateOTP() {
-				if (authConfig.DEFAULT_OTP) return authConfig.DEFAULT_OTP;
-				return Math.floor(100000 + Math.random() * 900000).toString();
-			},
-		}),
-		organization({
-			ac,
-			roles: orgRoles,
-			additionalFields: {
-				organization: {
-					billingEmail: {
-						type: "string",
-						required: false,
-						input: true,
-					},
-					billingName: {
-						type: "string",
-						required: false,
-						input: true,
-					},
-					externalCustomerId: {
-						type: "string",
-						required: false,
-						input: true,
-						unique: true,
-					},
-					status: {
-						type: "string",
-						required: false,
-						input: true,
-						defaultValue: "active",
-					},
-				},
-			},
-			async sendInvitationEmail(data) {
-				const inviteLink = `${authConfig.BASE_URL}/dashboard/signup?inviteId=${data.id}`;
-				log.info({
-					...{
-						email: data.email,
-						organization: data.organization.name,
-					},
-					message: "📧 Organization invitation email requested:",
-				});
-
-				// Log invite URL in development for easy testing
-				if (authConfig.NODE_ENV === "development") {
-					log.info("server", `🔗 Invite URL (DEV): ${inviteLink}`);
-				}
-
-				const isResend =
-					Date.now() - new Date(data.invitation.createdAt).getTime() > 5000;
-
-				try {
-					await bus.publish(
-						BusEvent.INVITE_CREATED,
-						{
-							email: data.email,
-							organizationName: data.organization.name,
-							inviteLink,
-							inviterName:
-								data.inviter.user.name ||
-								data.inviter.user.email.split("@")[0] ||
-								"Someone",
-							inviterEmail: data.inviter.user.email,
-							isResend,
-						},
-						{ msgId: `invite_created:${data.id}:${Date.now()}` },
-					);
-					log.info(
-						"server",
-						`✅ Organization invite bus event published for ${data.email} (resend: ${isResend})`,
-					);
-				} catch (error) {
-					log.error(
-						"server",
-						`❌ Failed to publish organization invite event:${error}`,
-					);
-				}
-			},
-			organizationHooks: {
-				afterAcceptInvitation: async ({ member, user, organization }) => {
-					try {
-						await bus.publish(
-							BusEvent.ORGANIZATION_JOINED,
-							{
-								organizationId: organization.id,
-								orgName: organization.name,
-								userId: user.id,
-								userEmail: user.email,
-								memberName: user.name || user.email,
-								role: member.role,
-								inviterName: "Admin",
-							},
-							{ msgId: `org_joined:${organization.id}:${user.id}` },
-						);
-						log.info(
-							"server",
-							`✅ Organization joined bus event published for ${user.email}`,
-						);
-
-						// Set active organization for the user
-						await db
-							.update(schema.user)
-							.set({ activeOrganizationId: organization.id })
-							.where(eq(schema.user.id, user.id));
-						log.info(
-							"server",
-							`✅ Active organization set to ${organization.id} for user ${user.id}`,
-						);
-					} catch (error) {
-						log.error({
-							...{ data: error },
-							message:
-								"❌ Failed to publish organization joined event or update user:",
-						});
-					}
-				},
-			},
-		}),
-		openAPI({
-			path: "/docs",
-		}),
-	],
-	advanced: {
-		cookiePrefix: "reloop",
-		ipAddress: {
-			ipAddressHeaders: ["x-client-ip", "x-forwarded-for"],
-		},
-	},
-});
-
-let _schema: ReturnType<typeof auth.api.generateOpenAPISchema> | null = null;
-
-const getSchema = async () => {
-	if (!_schema) {
-		_schema = auth.api.generateOpenAPISchema();
-	}
-	return _schema;
-};
-
-export const OpenAPI = {
-	getPaths: async (prefix = "/api/auth/v1") => {
-		try {
-			const { paths } = await getSchema();
-			const reference: Record<string, any> = {};
-
-			for (const path of Object.keys(paths)) {
-				const pathData = paths[path];
-				if (!pathData) continue;
-
-				const key = prefix + path;
-				reference[key] = { ...pathData };
-
-				for (const method of Object.keys(pathData)) {
-					const operation = reference[key][method];
-					if (operation && typeof operation === "object") {
-						operation.tags = ["Better Auth"];
-					}
-				}
-			}
-
-			return reference;
-		} catch (error) {
-			log.error({
-				...{ data: error },
-				message: "Failed to generate OpenAPI paths:",
-			});
-			return {};
-		}
-	},
-	components: async () => {
-		try {
-			const { components } = await getSchema();
-			return components;
-		} catch (error) {
-			log.error({
-				...{ data: error },
-				message: "Failed to generate OpenAPI components:",
-			});
-			return {};
-		}
-	},
-} as const;
+/**
+ * Thin re-export: the single Better Auth runtime lives in `@reloop/auth/server`.
+ */
+export { auth, OpenAPI } from "@reloop/auth/server";
