--- apps/backend/auth/src/lib/redis.ts
@@ -1,3 +1,0 @@
-import { RedisCache } from "@reloop/cache/redis-client";
-import { authConfig } from "../auth.config";
-export const redis = new RedisCache("auth", 30 * 60, authConfig.REDIS_URL);
