+++ packages/auth/src/server.ts
@@ -1,54 +1,5 @@
-import { apiKey } from "@better-auth/api-key";
-import { betterAuth } from "better-auth";
-import { admin, bearer, jwt, openAPI, organization } from "better-auth/plugins";
-import { ac, orgRoles } from "./permissions";
-import { platformAc, platformRoles } from "./platform-permissions";
-import { DEFAULT_USER_ROLE, PLATFORM_ADMIN_ROLE } from "./roles";
-
-export const auth = betterAuth({
-	baseURL: process.env.BASE_URL || "https://local.reloop.sh",
-	user: {
-		additionalFields: {
-			activeOrganizationId: {
-				type: "string",
-				required: false,
-				input: true,
-			},
-			mode: {
-				type: "string",
-				required: false,
-				input: true,
-				defaultValue: "dev",
-			},
-		},
-	},
-	basePath: "/api/auth/v1",
-	advanced: {
-		cookiePrefix: "reloop",
-	},
-	plugins: [
-		jwt(),
-		bearer(),
-		admin({
-			defaultRole: DEFAULT_USER_ROLE,
-			adminRoles: [PLATFORM_ADMIN_ROLE],
-			ac: platformAc,
-			roles: platformRoles,
-		}),
-		apiKey({ defaultPrefix: "rl" }),
-		organization({
-			ac,
-			roles: orgRoles,
-			sendInvitationEmail: async () => {},
-		}),
-		openAPI({ path: "/docs" }),
-	],
-});
-
-export type AuthInstance = typeof auth;
-export type User = typeof auth.$Infer.Session.user;
-export type Session = typeof auth.$Infer.Session & {
-	user: User & {
-		activeOrganizationId: string;
-	};
-};
+/**
+ * Package root re-export for `@reloop/auth/server`.
+ * Implementation lives under `./server/` so server-only modules stay grouped.
+ */
+export * from "./server/index";
