+++ packages/auth/src/server/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Server-only surface for `@reloop/auth`.
+ *
+ * Contains the single runtime Better Auth instance and helpers that depend on
+ * database / Redis / bus. Do not import this path from browser bundles.
+ */
+export { auth, OpenAPI } from "./auth";
+export { authServerConfig } from "./config";
+export { redis as authRedis } from "./redis";
+export * from "./signup-invite";
